adcionando colecao 2 na WebApi tbRecebimentoResumoManual

+++ api/Models/tbRecebimentoResumoManual.cs
@@ -9,8 +9,8 @@ namespace api.Models
         public Nullable<int> cdAdquirente { get; set; }
         public DateTime dtVenda { get; set; }
         public decimal vlVenda { get; set; }
-        public Nullable<int> qtTracacao { get; set; }
-        public int tpOperacao { get; set; }
+        public Nullable<byte> qtTracacao { get; set; }
+        public byte tpOperacao { get; set; }
         public Nullable<int> cdBandeira { get; set; }
         public virtual tbAdquirente tbAdquirente { get; set; }
         public virtual tbBandeira tbBandeira { get; set; }
